ref(pull): break up factories and fix some arroyo callsites.
