+++ rust_snuba/Cargo.toml
@@ -69,7 +69,7 @@ sentry = { version = "0.48.5", default-features = false, features = [
 sentry-kafka-schemas = "2.1.39"
 sentry-options = "1.2.1"
 sentry_protos = "0.60.0"
-sentry_arroyo = { git = "https://github.com/getsentry/arroyo", rev = "1bf5057", features = ["ssl"] }
+sentry_arroyo = { git = "https://github.com/getsentry/arroyo", branch = "rbroughan/pull-based-arroyo-part-2", features = ["ssl"] }
 sentry_usage_accountant = { version = "0.1.2", features = ["kafka"] }
 seq-macro = "0.3"
 serde = { version = "1.0", features = ["derive"] }
